requirements.txt
boto3
PyGithub
pyyaml